requirements.txt
numpy>=1.26.4
pandas>=2.2.2
matplotlib>=3.8.4
scikit-learn>=1.4.2
scikits.bootstrap==1.1.0
scipy>=1.13.0
scikit-learn>=1.4.2

